Made DAO folder and moved DAO classes
Just added a folder for easier navigation.
