+++ src/main/java/edu/wpi/teamg/DAOs/EdgeDAO.java
@@ -1,0 +1,4 @@
+package edu.wpi.teamg.DAOs;
+
+public interface EdgeDAO extends IEDAO {
+}
