+++ src/main/java/edu/wpi/teamg/DAOs/IEDAO.java
@@ -1,0 +1,6 @@
+package edu.wpi.teamg.DAOs;
+
+import edu.wpi.teamg.DAOs.DAO;
+
+public interface IEDAO extends DAO {
+}
