+++ src/main/java/edu/wpi/teamg/DAOs/LocationNameDAO.java
@@ -1,0 +1,6 @@
+package edu.wpi.teamg.DAOs;
+
+import edu.wpi.teamg.DAOs.IEDAO;
+
+public interface LocationNameDAO extends IEDAO {
+}
