+++ src/main/java/edu/wpi/teamg/DAOs/MoveDAO.java
@@ -1,0 +1,6 @@
+package edu.wpi.teamg.DAOs;
+
+import edu.wpi.teamg.DAOs.IEDAO;
+
+public interface MoveDAO extends IEDAO {
+}
