+++ src/main/java/edu/wpi/teamg/DAOs/RequestDAO.java
@@ -1,0 +1,7 @@
+package edu.wpi.teamg.DAOs;
+
+import edu.wpi.teamg.DAOs.IEDAO;
+
+public interface RequestDAO extends IEDAO {
+
+}
