--- src/main/java/edu/wpi/teamg/EdgeDAO.java
@@ -1,4 +1,0 @@
-package edu.wpi.teamg;
-
-public interface EdgeDAO extends IEDAO{
-}
