--- src/main/java/edu/wpi/teamg/IEDAO.java
@@ -1,4 +1,0 @@
-package edu.wpi.teamg;
-
-public interface IEDAO extends DAO{
-}
