--- src/main/java/edu/wpi/teamg/LocationNameDAO.java
@@ -1,4 +1,0 @@
-package edu.wpi.teamg;
-
-public interface LocationNameDAO extends IEDAO{
-}
