--- src/main/java/edu/wpi/teamg/MoveDAO.java
@@ -1,4 +1,0 @@
-package edu.wpi.teamg;
-
-public interface MoveDAO extends IEDAO {
-}
