--- src/main/java/edu/wpi/teamg/RequestDAO.java
@@ -1,5 +1,0 @@
-package edu.wpi.teamg;
-
-public interface RequestDAO extends IEDAO{
-
-}
